langchain_latest/simpler_rag/requirements.txt
langchain
langchain_community
langchain-google-genai
bs4
pypdf
chromadb
faiss-cpu
python_dotenv


streamlit